forge fmt

+++ eth-bridge/contracts/script/Deploy.s.sol
@@ -8,8 +8,8 @@ import "../src/Bridge.sol";
 
 contract Deploy is Script {
     function run() external {
-        uint256 fee = (101411 + 7 * 29958) * 11/uint256(10);
-        uint delay = 300; // * 12 sec = 1h;
+        uint256 fee = (101411 + 7 * 29958) * 11 / uint256(10);
+        uint256 delay = 300; // * 12 sec = 1h;
         uint32 votesRequired = 2;
 
         Bridge bridgeImpl = new Bridge();
